Move test modules out of the installable trusts package.
Relocate core and issue regression tests into the top-level tests/
Django app, put shared fixtures in tests/support.py, and discover the
suite by explicit module labels. The isolated custom-user runner is
unchanged. No migrates.md entry: installed behavior and schema do not
change.
