+++ docs/test-coverage.md
@@ -4,6 +4,17 @@ Recorded with the #14/#15 modernization. Existing tests were kept and
 extended. This is not a claim that the declarative permission model is
 complete.
 
+Executable tests live in the top-level `tests/` Django app, not the
+installable `trusts/` package:
+
+- `tests/support.py` — fixtures and helpers (not a `test*.py` module)
+- `tests/core/test_core.py` — ordinary authorization tests
+- `tests/regressions/test_issue_*.py` — historical issue regressions
+- `python -m tests.runtests` — discovers the core and regression modules
+  by explicit labels (`tests.core.test_core`, `tests.regressions.test_issue_*`)
+- `python -m tests.runtests_custom` — isolated custom-user suite
+  (`tests.custom_content`)
+
 ## Tests that must keep passing
 
 Denied access and organization isolation (a trust is the organization
