+++ tests/regressions/test_issue_26.py
@@ -33,8 +33,8 @@ from trusts.models import (
     TrustUserPermission,
     _resolve_configured_permission,
 )
-from trusts.test_issue8 import Issue8FixtureMixin
-from trusts.tests import enable_local_group_grant, reload_test_users
+from tests.regressions.test_issue_8 import Issue8FixtureMixin
+from tests.support import enable_local_group_grant, reload_test_users
 
 
 def _run_manage_py_check():
