+++ tests/runtests.py
@@ -11,11 +11,26 @@ from django.test.utils import get_runner
 from django.conf import settings
 
 
+# Explicit module labels so core and regression tests run without relying on
+# unittest's test*.py filename pattern. The isolated custom-user suite stays
+# on tests.runtests_custom (tests.custom_content) and is not included here.
+NORMAL_SUITE = [
+    'tests.core.test_core',
+    'tests.regressions.test_issue_4',
+    'tests.regressions.test_issue_8',
+    'tests.regressions.test_issue_23',
+    'tests.regressions.test_issue_25',
+    'tests.regressions.test_issue_26',
+    'tests.regressions.test_issue_29',
+    'tests.regressions.test_issue_33',
+]
+
+
 def runtests():
     django.setup()
     TestRunner = get_runner(settings)
     test_runner = TestRunner(verbosity=1, interactive=False)
-    failures = test_runner.run_tests(['trusts'])
+    failures = test_runner.run_tests(NORMAL_SUITE)
     sys.exit(bool(failures))
